added minor tweaks

+++ a3/src/main/java/csc435/moocme/a3/CourseServlet.java
@@ -76,69 +76,75 @@ public class CourseServlet extends HttpServlet {
 
         Connection conn = null;
         Statement stmt = null;
-        try {
-           Class.forName("com.mysql.jdbc.Driver");  
-           conn = DriverManager.getConnection(
-              "jdbc:mysql://localhost:3306/moocs?useSSL=false", "root", ""); 
-   
-           stmt = conn.createStatement();
+        if (req.getParameter("authenticated") != null && req.getParameter("authenticated").equals("true")) { 
+            try {
+            Class.forName("com.mysql.jdbc.Driver");  
+            conn = DriverManager.getConnection(
+                "jdbc:mysql://localhost:3306/moocs?useSSL=false", "root", ""); 
+    
+            stmt = conn.createStatement();
 
-           String sqlStr;
+            String sqlStr;
 
-           // jackson, parse json in body
-           ObjectMapper mapper = new ObjectMapper();
-           ReqJsonObject newObj = mapper.readValue(req.getInputStream(), ReqJsonObject.class);
+            // jackson, parse json in body
+            ObjectMapper mapper = new ObjectMapper();
+            ReqJsonObject newObj = mapper.readValue(req.getInputStream(), ReqJsonObject.class);
 
-           // get course id from uri
-           String[] resources = req.getRequestURI().substring(req.getContextPath().length()).split("/");
-           String courseId = resources[resources.length - 1];
+            // get course id from uri
+            String[] resources = req.getRequestURI().substring(req.getContextPath().length()).split("/");
+            String courseId = resources[resources.length - 1];
 
-           sqlStr = "update courses set title = \"" + newObj.title + "\"," + 
-                                        "institution = \"" + newObj.institution + "\"," +
-                                        "uri = \"" + newObj.url + "\"," +
-                                        "free = " + newObj.free + "," + 
-                                        "platform = \"" + newObj.platform + "\"" +
-                                        "where id=" + courseId + ";";
+            sqlStr = "update courses set title = \"" + newObj.title + "\"," + 
+                                            "institution = \"" + newObj.institution + "\"," +
+                                            "uri = \"" + newObj.url + "\"," +
+                                            "free = " + newObj.free + "," + 
+                                            "platform = \"" + newObj.platform + "\"" +
+                                            "where id=" + courseId + ";";
 
-           stmt.executeUpdate(sqlStr);
+            stmt.executeUpdate(sqlStr);
 
-           sqlStr = "select * from courses where id=" + courseId + ";";
+            sqlStr = "select * from courses where id=" + courseId + ";";
 
-           ReqJsonObject ret = null;
-           ResultSet rs = stmt.executeQuery(sqlStr);
+            ReqJsonObject ret = null;
+            ResultSet rs = stmt.executeQuery(sqlStr);
 
-           // make query results RJO jackson can jsonify
-           while (rs.next()) {
-               ret = new ReqJsonObject(rs.getInt("id"),
-                                       rs.getString("title"),
-                                       rs.getString("platform"),
-                                       rs.getString("institution"),
-                                       rs.getString("uri"),
-                                       rs.getBoolean("free"));
-           }
+            // make query results RJO jackson can jsonify
+            while (rs.next()) {
+                ret = new ReqJsonObject(rs.getInt("id"),
+                                        rs.getString("title"),
+                                        rs.getString("platform"),
+                                        rs.getString("institution"),
+                                        rs.getString("uri"),
+                                        rs.getBoolean("free"));
+            }
 
-           // return query results jsonified
-           mapper = new ObjectMapper();
-           String jsonString = mapper.writerWithDefaultPrettyPrinter().writeValueAsString(ret);
+            // return query results jsonified
+            mapper = new ObjectMapper();
+            String jsonString = mapper.writerWithDefaultPrettyPrinter().writeValueAsString(ret);
 
-           out.print(jsonString);
+            out.print(jsonString);
 
-           } catch (SQLException ex) {
-                ex.printStackTrace();
-                out.print("{ \"success\": false }");
-           } catch (ClassNotFoundException ex) {
-                ex.printStackTrace();
-                out.print("{ \"success\": false }");
-           } finally {
-                out.flush();
-                out.close();  
-                try {
-                    if (stmt != null) stmt.close();
-                    if (conn != null) conn.close();
-                } catch (SQLException ex) {
+            } catch (SQLException ex) {
                     ex.printStackTrace();
-                }
-           }
+                    out.print("{ \"success\": false }");
+            } catch (ClassNotFoundException ex) {
+                    ex.printStackTrace();
+                    out.print("{ \"success\": false }");
+            } finally {
+                    out.flush();
+                    out.close();  
+                    try {
+                        if (stmt != null) stmt.close();
+                        if (conn != null) conn.close();
+                    } catch (SQLException ex) {
+                        ex.printStackTrace();
+                    }
+            }
+        } else {
+            out.print("{ \"success\": false }");
+            out.flush();
+            out.close();    
+        }
     }
 
     @Override
@@ -149,42 +155,48 @@ public class CourseServlet extends HttpServlet {
 
         Connection conn = null;
         Statement stmt = null;
-        try {
-           Class.forName("com.mysql.jdbc.Driver");  
-           conn = DriverManager.getConnection(
-              "jdbc:mysql://localhost:3306/moocs?useSSL=false", "root", ""); 
-   
-           stmt = conn.createStatement();
+        if (req.getParameter("authenticated") != null && req.getParameter("authenticated").equals("true")) {
+            try {
+            Class.forName("com.mysql.jdbc.Driver");  
+            conn = DriverManager.getConnection(
+                "jdbc:mysql://localhost:3306/moocs?useSSL=false", "root", ""); 
+    
+            stmt = conn.createStatement();
 
-           String sqlStr;
+            String sqlStr;
 
-           // get course id to delete
-           String[] resources = req.getRequestURI().substring(req.getContextPath().length()).split("/");
-           String courseId = resources[resources.length - 1];
+            // get course id to delete
+            String[] resources = req.getRequestURI().substring(req.getContextPath().length()).split("/");
+            String courseId = resources[resources.length - 1];
 
-           sqlStr = "delete from courses where id=" + courseId + ";";
+            sqlStr = "delete from courses where id=" + courseId + ";";
 
-           stmt.executeUpdate(sqlStr);
+            stmt.executeUpdate(sqlStr);
 
-           sqlStr = "select * from courses where id=" + courseId + ";";
+            sqlStr = "select * from courses where id=" + courseId + ";";
 
-           out.print("{ \"success\": true }");
+            out.print("{ \"success\": true }");
 
-           } catch (SQLException ex) {
-                ex.printStackTrace();
-                out.print("{ \"success\": false }");
-           } catch (ClassNotFoundException ex) {
-                ex.printStackTrace();
-                out.print("{ \"success\": false }");
-           } finally {
-                out.flush();
-                out.close();  
-                try {
-                    if (stmt != null) stmt.close();
-                    if (conn != null) conn.close();
-                } catch (SQLException ex) {
+            } catch (SQLException ex) {
                     ex.printStackTrace();
-                }
-           }
+                    out.print("{ \"success\": false }");
+            } catch (ClassNotFoundException ex) {
+                    ex.printStackTrace();
+                    out.print("{ \"success\": false }");
+            } finally {
+                    out.flush();
+                    out.close();  
+                    try {
+                        if (stmt != null) stmt.close();
+                        if (conn != null) conn.close();
+                    } catch (SQLException ex) {
+                        ex.printStackTrace();
+                    }
+            }
+        } else {
+            out.print("{ \"success\": false }");
+            out.flush();
+            out.close();
+        }
     }
 }
